Renamed module "tasks" to "task"
